finally!!

+++ internal/models/credit.go
@@ -5,7 +5,7 @@ import "time"
 type Credit struct {
 	ID             int        `db:"id"`
 	UserID         int        `db:"user_id"`
-	Amount         float64    `db:"amount"`
+	Amount         int64      `db:"amount"`
 	Currency       string     `db:"currency"`
 	DurationMonths int        `db:"duration_months"`
 	InterestRate   float64    `db:"interest_rate"`
